Starting to debug large disk (greater than 250MB) support on the IDE controller.
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@687 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/hardware/ide/HdGeometry.java
@@ -41,7 +41,7 @@ public class HdGeometry extends Internal {
             return -1;
         }
         /* test msdos magic */
-        if (buf[510] != 0x55 || buf[511] != 0xaa) {
+        if (buf[510] != 0x55 || buf[511] != (byte)0xaa) {
             return -1;
         }
         for (i = 0; i < 4; i++) {
